Add fallback if credentials render nil
Tests are currently failing because travis is stuck on accessing a ruby variable
it doesn't have access to. This fix should hopefully allow the url to
concatenate even if the variable is nil.

External access has been blocked in tests so this url won't need to work in that
case.